+++ src/User/UserAddress.php
@@ -7,7 +7,9 @@ declare(strict_types = 1);
 
 namespace Wizaplace\SDK\User;
 
-final class UserAddress
+use Wizaplace\SDK\ArrayableInterface;
+
+final class UserAddress implements ArrayableInterface
 {
     /** @var null|UserTitle */
     private $title;
@@ -104,5 +106,21 @@ final class UserAddress
     public function getCountry(): string
     {
         return $this->country;
+    }
+
+    public function toArray() : array
+    {
+        return [
+            'title'     => is_null($this->getTitle()) ? null : $this->getTitle()->getValue(),
+            'firstname' => $this->getFirstName(),
+            'lastname'  => $this->getLastName(),
+            'company'   => $this->getCompany(),
+            'phone'     => $this->getPhone(),
+            'address'   => $this->getAddress(),
+            'address_2' => $this->getAddressSecondLine(),
+            'zipcode'   => $this->getZipCode(),
+            'city'      => $this->getCity(),
+            'country'   => $this->getCountry(),
+        ];
     }
 }