+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- Added optionals parameters `$billing` and `$shipping` to `\Wizaplace\SDK\Organisation\OrganisationService::register`
+
 ## 1.36.0
 
 ### New features